move NilContext
