+++ src/frames/frame-atom.ts
@@ -1,5 +1,5 @@
 import { type Any, Frame } from "./frame.ts";
-import { NilContext } from "./meta-frame.ts";
+import { NilContext } from "./context.ts";
 
 export class FrameAtom extends Frame {
   constructor(meta = NilContext) {
