+++ src/frames/frame-note.ts
@@ -3,7 +3,7 @@ import { FrameArray } from "./frame-array.ts";
 import { FrameQuote } from "./frame-atom.ts";
 import { FrameString } from "./frame-string.ts";
 import { FrameSymbol } from "./frame-symbol.ts";
-import { NilContext } from "./meta-frame.ts";
+import { NilContext } from "./context.ts";
 
 export type Binding = { [key: string]: string };
 export type LanguageBinding = { [key: string]: Binding };
