+++ src/frames/frame-schema.ts
@@ -1,7 +1,7 @@
 import { Frame } from "./frame.ts";
 import { FrameList } from "./frame-list.ts";
 import { FrameNote } from "./frame-note.ts";
-import { NilContext } from "./meta-frame.ts";
+import { NilContext } from "./context.ts";
 
 export class FrameSchema extends FrameList {
   public static readonly BEGIN_SCHEMA = "<";
