+++ src/frames/frame.ts
@@ -1,5 +1,5 @@
 import { MetaFrame } from "./meta-frame.ts";
-import { NilContext } from "./meta-frame.ts";
+import { NilContext } from "./context.ts";
 import type { ICurryFunction } from "../ops.ts";
 import type { IArrayConstructor } from "../frames.ts";
